speed_PRO/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib
pyyaml
aiohttp
ping3